error fixed

+++ app/Http/Controllers/Admin/PagesController.php
@@ -127,7 +127,7 @@ class PagesController extends Controller
                 'imageable_type' => 'App\Models\Page',
                 'user_id' => $user->id
             ])->first();
-            if($image->path){
+            if($image && $image->path){
                 UploadService::destroyFile($image->path);
             }
