+++ app/Http/Controllers/User/UsersController.php
@@ -84,7 +84,7 @@ class UsersController extends Controller
                 'imageable_type' => 'App\User',
                 'user_id' => $user->id
             ])->first();
-            if($image->path){
+            if($image && $image->path){
                 UploadService::destroyFile($image->path);
             }
